context-runners: sanitize non-string fields
Closes #641, closes #632
Fixes #630, fixes #580
