+++ src/context-runners/validate.ts
@@ -1,25 +1,12 @@
 import { Context } from '../context';
 import { ContextRunner, FieldInstance } from './context-runner';
 import { Request, ValidationError } from '../base';
+import { toString } from '../utils';
 
 class UnknownError extends Error {
   constructor() {
     super();
   }
-}
-
-export function toString(value: any, deep = true): string {
-  if (Array.isArray(value) && value.length && deep) {
-    return toString(value[0], false);
-  } else if (value instanceof Date) {
-    return value.toISOString();
-  } else if (value && typeof value === 'object' && value.toString) {
-    return value.toString();
-  } else if (value == null || (isNaN(value) && !value.length)) {
-    return '';
-  }
-
-  return String(value);
 }
 
 export class Validate implements ContextRunner {
