+++ src/utils.ts
@@ -9,3 +9,17 @@ export const bindAll = <T>(object: T): { [K in keyof T]: T[K] } => {
 
   return object;
 };
+
+export function toString(value: any, deep = true): string {
+  if (Array.isArray(value) && value.length && deep) {
+    return toString(value[0], false);
+  } else if (value instanceof Date) {
+    return value.toISOString();
+  } else if (value && typeof value === 'object' && value.toString) {
+    return value.toString();
+  } else if (value == null || (isNaN(value) && !value.length)) {
+    return '';
+  }
+
+  return String(value);
+}
